vm: prevent excessive memory usage due to stack pollution
If the return value of a function is not used in Daedalus code, it remains on the stack. Due to this, if the stack is not cleared regularly, it can fill up with unused return values which is basically a memory leak. By emptying the stack after each call into the script, this is avoided.