vm: add execution flags for configuring how the VM behaves
This needed to be implemented due to bugs present in both
Daedalus scripts and the original VM implementation which
cause accesses of variables on invalid instances to work
regardless.

Now, the `vm_allow_null_instance_access` execution flag
can be set which causes the VM to pop a `0` whenever such
invalid access is attempted instead of throwing an exception.