[#35] project: improve dependency handling and re-add git submodules
Due to issues in build environments without a network connection,
it is necessary to allow users to pull phoenix's dependencies before
running any build scripts. By re-adding git submodules for the
dependencies, this can be made possible.

Because working with git submodules on larger projects with many
nested dependencies is a pain, however, phoenix still supports
downloading dependencies while configuring, so using the submodule
approach is not required.