[#58] fix(script): fix function parameters
This commit fixes a crash with higher-order functions. Previously, passing a function as a parameter might cause the callee to copy its value into a local variable which did not have enough space for the function value.

With this fix, one integer of space is allocated for non-const function symbols which should fix that issue.