[#58] feat(script,vm): move const-ness check to the VM
Previously, the `symbol` did not allow for the mutation of symbols declared `const`. This presents an issue whenever the VM's owner wants to apply some sort of runtime fix or just generally wants to be able to mutate the VM entirely.

Changing this also allows for specifying a VM execution flag for allowing the mutation of symbols declared `const`. This is of use in some corner cases, like Gothic mods which exploit a vulnerability in the original VM implementation to gain direct memory access.