[#60] fix(archive_binary): fix incorrect whitespace ignore in header
Previously, the parser for binary archives would parse the last element of the ZEN header (which is "END") and proceed to skip whitespace after it.

This can cause the parser to read bytes of the content of the archive, not belonging to the header, should they start with `0x20`, `\t`, `\v`, `\f`, `\r` or '\n'. This fix introduces a customizable version of `buffer::get_line` which allows the caller to specify the byte values to skip.