refactor(Vfs): replace deprecated vdfs.hh with Vfs.hh
The VDF-implementation in `vdfs.hh` is very basic, very inflexible implementation which only supported loading VDF-files. The implementation completely ignored the original premise of the virtual file system in the ZenGin: unifying file system access.

The new implementation which can be found in `Vfs.hh` keeps this in mind and is designed from the ground up as a full virtual file system. It allows for mounting VDF-files, host file structures and more. Many of the unnecessary features of VDF-files (like catalog offset) have been removed so that the implementation can be slim and flexible.

Users of the old `vdfs.hh` implementation should migrate to the new `Vfs.hh`-implementation since the old implementation will be removed in phoenix 2.0