[#70] fix(buffer): when `mmap`-ing, don't fail on empty files
This patch introduces a fix to a long-standing problem with the buffer `mmap`-ing behaviour and empty files. Previously, when trying to map an empty file, the system would throw an exception which caused a bunch of problems in implementations. Now, it checks the file size before trying to map it and if the size is zero, the empty buffer is returned.

Closes: #70
(cherry picked from commit ed37464ca1cbe55d3ce708b0454652c3bea89aec)