[#70] fix(buffer): when reading a line at eof, return ""
This patch adds a check to `get_line_and_ignore` which causes the function to return the empty string if the buffer is at EOF. This fixes an issue with some Gothic mod files which don't contain a source string in their source sections.

Correction ed37464: that patch did not fix the issue mentioned in #70.