[#69] fix(light_preset): refine implementation
This patch aims to improve compatibility with the original implementation. It changes the `colorAniList` parser of `light_preset` to correctly handle greyscale color. In addition, it changes the alpha-value reported as part of the light color to `0xff` like the original implementation does.

Reference the `zCVobLightData::Unarchive` member function at `0x00609fa0` in the original binary.