refactor(Archive): initial implementation of spec-compliance
This patch adapts the implementation of `ReadArchive` to properly support object references in archives.

To do this, I added a new `Object` class which represents any ZenGin object and a new `read_object` API to `ReadArchive` which facilitates loading of a set of pre-defined ZenGin object types. This new API takes care of creating new object instances and resolving archive-internal references.

This required changing the `VirtualObject` hierarchy to use `shared_ptr` instead of `unique_ptr` so the system is able to share references to archive objects without sacrificing memory safety.