fix(Daedalus): fix member registration for G2 classic
G2 exclusive class members for `C_NPC` and `C_ITEM` are currently
registered based on a check if `effect` exists as member.

Problem is this is the only member that didn't exist in G2 classic but
was later introduced in the Addon. Add a separate check for `effect` to
have other members properly registered in case G2 classic is used.

Reported as bug in https://github.com/Try/OpenGothic/issues/840.